Correcao de bugs de requisitos ja feitos que nao estavam passando nos testes. As chamadas da primeira API de comidas/bebidas, estao sendo feitas no Provider agora.
