+++ src/context/Provider.js
@@ -1,6 +1,10 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import Context from '.';
+import fetchMeals from '../services/fetchMeals';
+import fetchMealsCategories from '../services/fetchMealsCategories';
+import fetchDrinks from '../services/fetchDrinks';
+import fetchDrinksCategories from '../services/fetchDrinksCategories';
 
 function Provider({ children }) {
   const [recipes, setRecipes] = useState([]);
@@ -14,6 +18,34 @@ function Provider({ children }) {
   const [filteredDrinks, setFilteredDrinks] = useState([]);
 
   const [toggleOn, setToggleOn] = useState('');
+
+  useEffect(() => {
+    const getMeals = async () => {
+      const results = await fetchMeals();
+      setMeals(results);
+      setFilteredMeals(results);
+    };
+    getMeals();
+
+    const getMealsCategories = async () => {
+      const results = await fetchMealsCategories();
+      setMealsCategories(results);
+    };
+    getMealsCategories();
+
+    const getDrinks = async () => {
+      const results = await fetchDrinks();
+      setDrinks(results);
+      setFilteredDrinks(results);
+    };
+    getDrinks();
+
+    const getDrinksCategories = async () => {
+      const results = await fetchDrinksCategories();
+      setDrinksCategories(results);
+    };
+    getDrinksCategories();
+  }, []);
 
   const context = {
     meals,
